requirements.txt
fastapi
uvicorn
openai
pinecone
PyMuPDF
langchain
python-dotenv
